scripts/naca0012_rans_only_with_export.py
#!/usr/bin/env python3
"""
naca0012_rans_only_with_export.py
=================================
Steady-state RANS sweep of a NACA 0012 aerofoil using the Spalart–Allmaras
turbulence model via the ANSYS PyFluent API.

Performs a full angle-of-attack sweep (0° → 22° → 0°) in 2° increments,
recording C_L and C_D at each angle. The script implements:

  - Automatic mesh validation (minimum 50,000 cells, y⁺ ≤ 1.5)
  - Moving-window convergence detection (σ(C_L), σ(C_D) < 10⁻⁴)
  - JSON-based checkpointing for automatic recovery from interruptions
  - Comprehensive CSV export of all aerodynamic coefficients

References
----------
  Ladson, C.L. (1988). NASA TM 4074.
  Menter, F.R. and Lechner, R. (2021). Best Practice: RANS Turbulence
    Modelling in Ansys CFD. Ansys Germany GmbH.

Usage
-----
  1. Launch Fluent with the mesh loaded:
       fluent 2ddp -t4 -g
  2. Run this script:
       python naca0012_rans_only_with_export.py

Author: [Your Name]
Date: 2026
Software: ANSYS Fluent 2025 R2 (Student), Python 3.10+
"""

import json
import math
import os
import sys
import time
from pathlib import Path

import numpy as np
import pandas as pd

try:
    import ansys.fluent.core as pyfluent
except ImportError:
    print("ERROR: ansys-fluent-core is not installed.")
    print("Install with: pip install ansys-fluent-core")
    sys.exit(1)

from utils.convergence_monitor import ConvergenceMonitor
from utils.vector_calculator import compute_lift_drag_vectors
from utils.checkpoint_manager import CheckpointManager

# =============================================================================
# Configuration
# =============================================================================

# Flow conditions
U_INF = 88.65          # Freestream velocity [m/s]
CHORD = 1.0            # Chord length [m]
RHO = 1.225            # Density [kg/m³]
MU = 1.81e-5           # Dynamic viscosity [Pa·s]
RE = RHO * U_INF * CHORD / MU  # Reynolds number (~6.0 × 10⁶)

# Turbulence inlet conditions
TURB_INTENSITY = 0.001  # 0.1%
TURB_VISC_RATIO = 1.0

# Sweep configuration
ANGLES_UP = list(range(0, 24, 2))          # 0° to 22° in 2° steps
ANGLES_DOWN = list(range(22, -2, -2))      # 22° to 0° in 2° steps

# Convergence criteria
CONVERGENCE_WINDOW = 50       # Iterations per sampling window
CONVERGENCE_THRESHOLD = 1e-4  # Standard deviation threshold for C_L, C_D
MIN_ITERATIONS = 100          # Minimum iterations before checking convergence
MAX_ITERATIONS = 500          # Timeout for inherently unsteady angles

# Mesh validation thresholds
MIN_CELLS = 50_000
MAX_YPLUS = 1.5

# Output paths
OUTPUT_DIR = Path("../results")
CHECKPOINT_FILE = Path("../results/rans_checkpoint.json")
EXPORT_CSV = Path("../results/sa_rans_sweep.csv")


# =============================================================================
# Mesh validation
# =============================================================================

def validate_mesh(solver):
    """
    Check that the mesh meets minimum quality requirements.

    Parameters
    ----------
    solver : pyfluent.Session
        Active Fluent session.

    Raises
    ------
    RuntimeError
        If mesh fails validation checks.
    """
    # Check cell count
    mesh_info = solver.tui.mesh.check()
    cell_count = solver.setup.general.solver.domain_cell_count
    if cell_count < MIN_CELLS:
        raise RuntimeError(
            f"Mesh has {cell_count} cells, minimum required is {MIN_CELLS}."
        )
    print(f"  Mesh validation: {cell_count} cells — OK")

    # Note: y⁺ can only be checked after running at least a few iterations.
    # Post-run y⁺ verification is performed in the main loop.


# =============================================================================
# Solver setup
# =============================================================================

def setup_solver(solver):
    """
    Configure the Fluent solver for steady-state SA RANS.

    Parameters
    ----------
    solver : pyfluent.Session
        Active Fluent session.
    """
    # General settings
    general = solver.setup.general
    general.solver.type = "pressure-based"
    general.solver.time = "steady"
    general.solver.velocity_formulation = "absolute"

    # Turbulence model: Spalart–Allmaras
    viscous = solver.setup.models.viscous
    viscous.model = "spalart-allmaras"

    # Pressure-velocity coupling: Coupled
    methods = solver.solution.methods
    methods.p_v_coupling.scheme = "coupled"

    # Spatial discretisation
    methods.discretization.pressure = "second-order"
    methods.discretization.momentum = "second-order-upwind"
    methods.discretization.modified_turbulent_viscosity = "second-order-upwind"
    methods.gradient_scheme = "least-squares-cell-based"

    print("  Solver configured: Pressure-based, Steady, SA, Coupled")


def set_boundary_conditions(solver, alpha_deg):
    """
    Update velocity inlet direction and monitoring vectors for the given
    angle of attack.

    Parameters
    ----------
    solver : pyfluent.Session
        Active Fluent session.
    alpha_deg : float
        Angle of attack in degrees.
    """
    alpha_rad = math.radians(alpha_deg)
    u_x = U_INF * math.cos(alpha_rad)
    u_y = U_INF * math.sin(alpha_rad)

    # Update velocity inlet
    inlet = solver.setup.boundary_conditions.velocity_inlet["inlet"]
    inlet.velocity.value = U_INF
    inlet.flow_direction.x = math.cos(alpha_rad)
    inlet.flow_direction.y = math.sin(alpha_rad)

    # Update turbulence
    inlet.turbulent_intensity = TURB_INTENSITY
    inlet.turbulent_viscosity_ratio = TURB_VISC_RATIO

    # Update lift and drag monitoring directions
    lift_dir, drag_dir = compute_lift_drag_vectors(alpha_deg)
    solver.solution.report_definitions.force["cl"] = {
        "force_vector": lift_dir,
    }
    solver.solution.report_definitions.force["cd"] = {
        "force_vector": drag_dir,
    }

    print(f"  BCs set: α = {alpha_deg}°, u = ({u_x:.2f}, {u_y:.2f}) m/s")


# =============================================================================
# Main sweep loop
# =============================================================================

def run_sweep(solver, angles, direction, results, checkpoint_mgr):
    """
    Execute the angle-of-attack sweep.

    Parameters
    ----------
    solver : pyfluent.Session
        Active Fluent session.
    angles : list[int]
        Angles to simulate.
    direction : str
        'Pitch-Up' or 'Pitch-Down'.
    results : list[dict]
        Accumulator for results rows.
    checkpoint_mgr : CheckpointManager
        Checkpoint handler.
    """
    monitor = ConvergenceMonitor(
        window_size=CONVERGENCE_WINDOW,
        threshold=CONVERGENCE_THRESHOLD,
    )

    for alpha in angles:
        # Check if already completed (from checkpoint)
        if checkpoint_mgr.is_completed(alpha, direction):
            print(f"  Skipping α = {alpha}° ({direction}) — already completed")
            continue

        print(f"\n{'='*60}")
        print(f"  Simulating α = {alpha}° ({direction})")
        print(f"{'='*60}")

        set_boundary_conditions(solver, alpha)
        monitor.reset()

        # Run iterations with convergence monitoring
        converged = False
        total_iters = 0

        while total_iters < MAX_ITERATIONS:
            solver.solution.run_calculation.iterate(
                number_of_iterations=CONVERGENCE_WINDOW
            )
            total_iters += CONVERGENCE_WINDOW

            # Sample current coefficients
            cl = solver.solution.report_definitions.force["cl"].get_result()
            cd = solver.solution.report_definitions.force["cd"].get_result()
            monitor.update(cl, cd)

            if total_iters >= MIN_ITERATIONS and monitor.is_converged():
                converged = True
                print(f"  Converged at {total_iters} iterations: "
                      f"C_L = {cl:.5f}, C_D = {cd:.6f}")
                break

        if not converged:
            print(f"  Timeout at {total_iters} iterations (inherent unsteadiness): "
                  f"C_L = {cl:.5f}, C_D = {cd:.6f}")

        # Record result
        results.append({
            "alpha_deg": alpha,
            "direction": direction,
            "C_L": cl,
            "C_D": cd,
            "iterations": total_iters,
            "converged": converged,
        })

        # Checkpoint
        checkpoint_mgr.save(alpha, direction, cl, cd)

        # Export case/data at this angle
        solver.file.write(
            case_file=f"../results/sa_rans_alpha{alpha:02d}_{direction.lower().replace('-','')}.cas.h5",
            data_file=f"../results/sa_rans_alpha{alpha:02d}_{direction.lower().replace('-','')}.dat.h5",
        )

    return results


# =============================================================================
# Entry point
# =============================================================================

def main():
    """Main execution function."""
    print("=" * 60)
    print("  NACA 0012 — Spalart–Allmaras Steady RANS Sweep")
    print(f"  Re = {RE:.2e}")
    print("=" * 60)

    # Create output directory
    OUTPUT_DIR.mkdir(parents=True, exist_ok=True)

    # Connect to running Fluent session
    print("\nConnecting to Fluent session...")
    solver = pyfluent.connect_to_fluent()
    print("  Connected.")

    # Validate mesh
    print("\nValidating mesh...")
    validate_mesh(solver)

    # Configure solver
    print("\nConfiguring solver...")
    setup_solver(solver)

    # Initialise checkpoint manager
    checkpoint_mgr = CheckpointManager(CHECKPOINT_FILE)

    # Hybrid initialisation
    print("\nPerforming hybrid initialisation...")
    solver.solution.initialization.hybrid_initialize()
    print("  Hybrid initialisation complete.")

    # Run sweeps
    results = []
    print("\n" + "=" * 60)
    print("  PITCH-UP SWEEP: 0° → 22°")
    print("=" * 60)
    results = run_sweep(solver, ANGLES_UP, "Pitch-Up", results, checkpoint_mgr)

    print("\n" + "=" * 60)
    print("  PITCH-DOWN SWEEP: 22° → 0°")
    print("=" * 60)
    results = run_sweep(solver, ANGLES_DOWN, "Pitch-Down", results, checkpoint_mgr)

    # Export results
    df = pd.DataFrame(results)
    df.to_csv(EXPORT_CSV, index=False)
    print(f"\nResults exported to {EXPORT_CSV}")
    print(f"Total angles simulated: {len(results)}")

    # Summary
    print("\n" + "=" * 60)
    print("  SUMMARY")
    print("=" * 60)
    print(df.to_string(index=False))


if __name__ == "__main__":
    main()
